Separate vnf_p7.c into common, LTE and NR implementations
